refactor: folder structure, extract comp & style
