+++ resources/assets/js/page/Study/Filter/ItemFilter/ItemFilterColumn/style.js
@@ -1,9 +1,9 @@
 import { makeStyles } from '@material-ui/core/styles';
-import styled from 'styled-components';
-import { color } from '~/theme/global';
 import CategoryIcon from '@material-ui/icons/Category';
 import TimelineIcon from '@material-ui/icons/Timeline';
 import AccessAlarmIcon from '@material-ui/icons/AccessAlarm';
+import styled from 'styled-components';
+import { color } from '~/theme/global';
 
 export const ItemFilterColContainer = styled.div`
     width: 115px;
