+++ resources/assets/js/page/Study/Filter/ItemFilter/index.js
@@ -1,15 +1,9 @@
 import React, { forwardRef } from 'react';
 import map from 'lodash/map';
-import styled from 'styled-components';
 
-import ItemFilterColumn from '../ItemFilterColumn';
+import { ItemFilterContainer } from './style';
+import ItemFilterColumn from './ItemFilterColumn';
 import useItemFilter from '../useItemFilter';
-
-const ItemFilterContainer = styled.div`
-    display: flex;
-    flex-direction: row;
-    padding: 10px 0;
-`;
 
 const ItemFilter = forwardRef(() => {
     const filterObjArr = useItemFilter();
