+++ resources/assets/js/page/Study/Filter/ItemFilter/style.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const ItemFilterContainer = styled.div`
+    display: flex;
+    flex-direction: row;
+    padding: 10px 0;
+`;
